Added tenant processor
This first implementation of the tenant processor is able to extract bearer tokens from 'authorization' headers and parse the JWT, extracting the tenant from a field in the payload. This has been tested with the OTLP exporter's 'PerRPCAuth' option and a Kubernetes token from local disk (simulating what a pod would have access).

The processor adds a 'tenant' resource ('process tag' in Jaeger jargon) with the contents of the parsed tenant value to all spans going through the processor.

Signed-off-by: Juraci Paixão Kröhling <juraci@kroehling.de>